Harden blueprint update and metadata handling and refactor test suite
