+++ custom_components/blueprints_updater/translations/de.json
@@ -124,7 +124,7 @@
     "content_missing": "Der Blueprint-Inhalt fehlt oder konnte nicht abgerufen werden.",
     "git_diff_title": "Codeänderungen (Git Diff)",
     "semantic_sync_notice": "Dieser Blueprint ist inhaltlich unverändert, erfordert jedoch eine strukturelle Aktualisierung zur Anpassung an die neue kanonische Formatierung (z. B. Normalisierung der Zeilenenden, Metadaten-Injektion).",
-    "breaking_risks_title": "AKTUALISIERUNGSRISIKO-BERICHT",
+    "breaking_risks_title": "⚠️ POTENZIELLE BREAKING CHANGES",
     "auto_update_blocked_by_breaking_change": "Automatisches Update für {name} wegen erkannter schwerwiegender Änderungen blockiert",
     "risk_new_mandatory": "Neue obligatorische Eingabe {input} hinzugefügt.",
     "risk_missing_input": "Der Entität {entity} fehlt die für die neue Version erforderliche obligatorische Eingabe {input}.",
