+++ custom_components/blueprints_updater/translations/es.json
@@ -124,7 +124,7 @@
     "content_missing": "El contenido del blueprint falta o no se pudo obtener.",
     "git_diff_title": "Cambios en el código (Git Diff)",
     "semantic_sync_notice": "Este blueprint es semánticamente idéntico, pero requiere una actualización estructural para alinearse con el nuevo formato canónico (p. ej., normalización de saltos de línea, inyección de metadatos).",
-    "breaking_risks_title": "INFORME DE RIESGOS DE ACTUALIZACIÓN",
+    "breaking_risks_title": "⚠️ POSIBLES CAMBIOS DE COMPATIBILIDAD",
     "auto_update_blocked_by_breaking_change": "Actualización automática bloqueada para {name} debido a cambios importantes detectados",
     "risk_new_mandatory": "Nueva entrada obligatoria {input} añadida.",
     "risk_missing_input": "A la entidad {entity} le falta la entrada obligatoria {input} requerida por la nueva versión.",
