+++ custom_components/blueprints_updater/translations/ru.json
@@ -124,7 +124,7 @@
     "content_missing": "Содержимое чертежа отсутствует или не может быть получено.",
     "git_diff_title": "Изменения в коде (Git Diff)",
     "semantic_sync_notice": "Этот блюпринт логически не изменился, но требует структурного обновления для соответствия новому каноническому форматированию (например, нормализация переносов строк, внедрение метаданных).",
-    "breaking_risks_title": "ОТЧЕТ О РИСКАХ ОБНОВЛЕНИЯ",
+    "breaking_risks_title": "⚠️ ВОЗМОЖНЫЕ КРИТИЧЕСКИЕ ИЗМЕНЕНИЯ",
     "auto_update_blocked_by_breaking_change": "Автоматическое обновление {name} заблокировано из-за обнаружения критических изменений",
     "risk_new_mandatory": "Добавлен новый обязательный ввод {input}.",
     "risk_missing_input": "В сущности {entity} отсутствует обязательный ввод {input}, необходимый для новой версии.",
