+++ custom_components/blueprints_updater/translations/vi.json
@@ -124,7 +124,7 @@
     "content_missing": "Nội dung blueprint bị thiếu hoặc không thể tải xuống",
     "git_diff_title": "Mã nguồn thay đổi (Git Diff)",
     "semantic_sync_notice": "Bản blueprint này giống hệt về mặt logic, nhưng cần cập nhật cấu trúc để đồng bộ với định dạng chuẩn mới (ví dụ: chuẩn hóa xuống dòng, bổ sung metadata).",
-    "breaking_risks_title": "BÁO CÁO RỦI RO CẬP NHẬT",
+    "breaking_risks_title": "⚠️ CẢNH BÁO THAY ĐỔI LỚN",
     "auto_update_blocked_by_breaking_change": "Tự động cập nhật cho {name} bị chặn do phát hiện thay đổi lớn",
     "risk_new_mandatory": "Tham số bắt buộc mới {input} vừa được thêm vào.",
     "risk_missing_input": "Thực thể {entity} đang thiếu tham số bắt buộc {input} mà phiên bản mới yêu cầu.",
