+++ pyproject.toml
@@ -15,6 +15,7 @@ dev = [
   "pytest>=9",
   "pytest-asyncio",
   "pytest-cov",
+  "pytest-timeout",
   "ruff",
   "tomlkit",
   "ty",
@@ -50,7 +51,7 @@ indent-style = "space"
 asyncio_mode = "auto"
 pythonpath = ["."]
 filterwarnings = ["ignore::DeprecationWarning"]
-addopts = ["--cov=custom_components", "--cov-fail-under=90"]
+addopts = ["--cov=custom_components", "--cov-fail-under=90", "--timeout=60"]
 
 [tool.pyright]
 typeCheckingMode = "standard"
