+++ tests/coordinator/test_cache_clearing.py
