+++ tests/coordinator/test_compatibility_guard.py
