+++ tests/coordinator/test_schema_guard.py
