+++ tests/coordinator/test_storage.py
@@ -1,0 +1,473 @@
+"""Tests for coordinator storage, persistence, and metadata pruning."""
+
+import asyncio
+import os
+from types import MappingProxyType
+from unittest.mock import AsyncMock, MagicMock, patch
+
+import pytest
+
+from custom_components.blueprints_updater.coordinator import BlueprintUpdateCoordinator
+
+
+@pytest.mark.asyncio
+async def test_prune_preserves_hashes_only_metadata(coordinator):
+    """Test that blueprints with only a hash (and no ETag) are not pruned if they exist."""
+    path = "/config/blueprints/automation/hash_only.yaml"
+    coordinator._persisted_etags = {}
+    coordinator._persisted_hashes = {path: "some_hash"}
+
+    with (
+        patch("custom_components.blueprints_updater.coordinator.os.path.isfile", return_value=True),
+        patch.object(coordinator, "_async_save_metadata") as mock_save,
+    ):
+        await coordinator._async_prune_stale_metadata(set())
+
+    assert path in coordinator._persisted_hashes
+    assert coordinator._persisted_hashes[path] == "some_hash"
+    mock_save.assert_not_called()
+
+
+@pytest.mark.asyncio
+async def test_async_prune_stale_metadata_triggers_save(coordinator):
+    """Test that pruning stale metadata triggers a background save operation."""
+    path = "/config/blueprints/automation/stale.yaml"
+    coordinator._persisted_hashes = {path: "some_hash"}
+
+    with (
+        patch(
+            "custom_components.blueprints_updater.coordinator.os.path.isfile", return_value=False
+        ),
+        patch.object(coordinator, "_async_save_metadata") as mock_save,
+    ):
+        await coordinator._async_prune_stale_metadata(set())
+
+    assert path not in coordinator._persisted_hashes
+    mock_save.assert_called_once_with(force=True)
+
+
+@pytest.mark.asyncio
+async def test_save_metadata_honors_cleared_in_memory_state(coordinator):
+    """Test that clearing an ETag in-memory correctly results in it being removed from save."""
+    path = "/config/blueprints/automation/test.yaml"
+    coordinator._persisted_etags = {path: "old_etag"}
+    coordinator._persisted_hashes = {path: "old_hash"}
+
+    coordinator.data = {
+        path: {
+            "name": "Test",
+            "rel_path": "automation/test.yaml",
+            "source_url": "https://url.com",
+            "etag": None,
+            "remote_hash": None,
+        }
+    }
+
+    with (
+        patch.object(coordinator._store, "async_save") as mock_save,
+        patch("custom_components.blueprints_updater.coordinator.os.path.isfile", return_value=True),
+    ):
+        await coordinator._async_save_metadata(force=True, skip_filter=True)
+
+        mock_save.assert_called_once()
+        save_data = mock_save.call_args[0][0]
+        saved_etags = save_data["etags"]
+        saved_hashes = save_data["remote_hashes"]
+        assert path not in saved_etags
+        assert path not in saved_hashes
+
+
+@pytest.mark.asyncio
+async def test_metadata_pruning(coordinator):
+    """Test that stale metadata is pruned during update."""
+    path_valid = "/config/blueprints/automation/valid.yaml"
+    path_stale = "/config/blueprints/automation/stale.yaml"
+
+    coordinator._persisted_etags = {path_valid: "etag1", path_stale: "etag2"}
+    coordinator._persisted_hashes = {path_valid: "hash1", path_stale: "hash2"}
+
+    blueprints = {
+        path_valid: {
+            "name": "Valid",
+            "rel_path": "automation/valid.yaml",
+            "domain": "automation",
+            "source_url": "https://url",
+            "local_hash": "hash1",
+        }
+    }
+
+    with (
+        patch.object(coordinator, "scan_blueprints", return_value=blueprints),
+        patch.object(coordinator, "_start_background_refresh"),
+    ):
+        await coordinator._async_update_data()
+
+    assert path_valid in coordinator._persisted_etags
+    assert path_stale not in coordinator._persisted_etags
+    assert path_valid in coordinator._persisted_hashes
+    assert path_stale not in coordinator._persisted_hashes
+
+
+@pytest.mark.asyncio
+async def test_async_save_metadata_empty_data(coordinator):
+    """Test that saving metadata with empty data clears the store."""
+    coordinator.data = {}
+    coordinator.setup_complete = True
+    coordinator._persisted_etags = {"stale": "etag"}
+    coordinator._persisted_hashes = {"stale": "hash"}
+
+    with patch.object(coordinator._store, "async_save", new_callable=AsyncMock) as mock_save:
+        await coordinator._async_save_metadata()
+
+    mock_save.assert_called_once_with({"etags": {}, "remote_hashes": {}})
+    assert not coordinator._persisted_etags
+    assert not coordinator._persisted_hashes
+
+
+@pytest.mark.asyncio
+async def test_prune_metadata_persistence(coordinator):
+    """Test that stale metadata is pruned from memory and persisted to disk."""
+    path_exist = "/config/blueprints/automation/exist.yaml"
+    path_stale = "/config/blueprints/automation/stale.yaml"
+
+    coordinator._persisted_etags = {path_exist: "e1", path_stale: "e2"}
+    coordinator._persisted_hashes = {path_exist: "h1", path_stale: "h2"}
+    coordinator.setup_complete = True
+
+    coordinator.data = {path_exist: {"etag": "e1", "remote_hash": "h1"}}
+
+    tasks = []
+
+    def create_background_task(coro, name=None):
+        task = asyncio.create_task(coro)
+        tasks.append(task)
+        return task
+
+    with (
+        patch.object(
+            coordinator.hass, "async_create_background_task", side_effect=create_background_task
+        ),
+        patch.object(coordinator._store, "async_save", new_callable=AsyncMock) as mock_save,
+        patch(
+            "custom_components.blueprints_updater.coordinator.os.path.isfile",
+            side_effect=lambda p: p in {path_exist},
+        ),
+    ):
+        await coordinator._async_prune_stale_metadata({path_exist})
+
+        if tasks:
+            await asyncio.gather(*tasks)
+
+    assert path_stale not in coordinator._persisted_etags
+    assert path_stale not in coordinator._persisted_hashes
+    assert path_exist in coordinator._persisted_etags
+    assert mock_save.called, "async_save was not called"
+    saved_data = mock_save.call_args[0][0]
+    assert path_stale not in saved_data["etags"]
+    assert saved_data["etags"][path_exist] == "e1"
+    assert path_stale not in saved_data["remote_hashes"]
+    assert saved_data["remote_hashes"][path_exist] == "h1"
+
+
+@pytest.mark.asyncio
+async def test_cold_start_rehydration(coordinator):
+    """Test that persisted hashes are used on reboot but verified later."""
+    path = "/config/blueprints/automation/test.yaml"
+    local_hash = "current_hash"
+    coordinator.data = {}
+    coordinator._persisted_hashes = {path: local_hash}
+    coordinator._first_update_done = False
+
+    blueprints = {
+        path: {
+            "name": "Test",
+            "rel_path": "automation/test.yaml",
+            "domain": "automation",
+            "source_url": "https://url",
+            "local_hash": local_hash,
+        }
+    }
+
+    with (
+        patch.object(coordinator, "scan_blueprints", return_value=blueprints),
+        patch.object(coordinator, "_start_background_refresh"),
+    ):
+        results = await coordinator._async_update_data()
+
+    assert not results[path]["updatable"]
+    assert results[path]["remote_hash"] == local_hash
+    assert coordinator._first_update_done
+
+
+@pytest.mark.asyncio
+async def test_etag_invalidation_on_mismatch(coordinator):
+    """Test that ETag is invalidated when local and remote hashes mismatch on startup."""
+    path = "/config/blueprints/automation/test.yaml"
+    local_hash = "current_hash"
+    remote_hash = "stale_hash"
+    coordinator.data = {}
+    coordinator._persisted_hashes = {path: remote_hash}
+    coordinator._persisted_etags = {path: "stale_etag"}
+    coordinator._first_update_done = False
+
+    blueprints = {
+        path: {
+            "name": "Test",
+            "rel_path": "automation/test.yaml",
+            "domain": "automation",
+            "source_url": "https://url",
+            "local_hash": local_hash,
+        }
+    }
+
+    with (
+        patch.object(coordinator, "scan_blueprints", return_value=blueprints),
+        patch.object(coordinator, "_start_background_refresh"),
+    ):
+        results = await coordinator._async_update_data()
+
+    assert results[path]["updatable"]
+    assert results[path]["etag"] is None
+    assert results[path]["remote_hash"] == remote_hash
+
+
+@pytest.mark.asyncio
+async def test_persisted_metadata_not_reused_after_first_update(coordinator):
+    """Test that persisted hashes/ETags are only used for the very first update."""
+    path = "/config/blueprints/automation/test.yaml"
+    initial_hash = "initial_hash"
+    coordinator._persisted_hashes = {path: initial_hash}
+    coordinator._persisted_etags = {path: "initial_etag"}
+    coordinator._first_update_done = False
+
+    blueprints = {
+        path: {
+            "name": "Test",
+            "rel_path": "automation/test.yaml",
+            "domain": "automation",
+            "source_url": "https://url",
+            "local_hash": initial_hash,
+        }
+    }
+
+    with (
+        patch.object(coordinator, "scan_blueprints", return_value=blueprints),
+        patch.object(coordinator, "_start_background_refresh"),
+    ):
+        first_results = await coordinator._async_update_data()
+    coordinator.data = first_results
+    assert coordinator._first_update_done
+
+    coordinator._persisted_hashes[path] = "stale_hash"
+    coordinator._persisted_etags[path] = "stale_etag"
+
+    with (
+        patch.object(coordinator, "scan_blueprints", return_value=blueprints),
+        patch.object(coordinator, "_start_background_refresh"),
+    ):
+        results = await coordinator._async_update_data()
+
+    assert results[path]["remote_hash"] == initial_hash
+    assert results[path]["etag"] == "initial_etag"
+
+
+@pytest.mark.asyncio
+async def test_metadata_preservation_during_scan(coordinator):
+    """Test that existing metadata is preserved during a scan until refreshed."""
+    path = "/config/blueprints/automation/test.yaml"
+    local_hash = "some_hash"
+    coordinator.data = {
+        path: {
+            "local_hash": local_hash,
+            "remote_hash": "remote_hash",
+            "remote_content": "remote_content",
+            "updatable": False,
+            "invalid_remote_hash": "stale_error",
+            "last_error": "failed_previously",
+            "etag": "some_etag",
+        }
+    }
+
+    blueprints = {
+        path: {
+            "name": "Test",
+            "rel_path": "automation/test.yaml",
+            "domain": "automation",
+            "source_url": "https://url",
+            "local_hash": local_hash,
+        }
+    }
+
+    with (
+        patch.object(coordinator, "scan_blueprints", return_value=blueprints),
+        patch.object(coordinator, "_start_background_refresh"),
+    ):
+        results = await coordinator._async_update_data()
+
+    assert results[path]["invalid_remote_hash"] == "stale_error"
+    assert results[path]["last_error"] == "failed_previously"
+    assert results[path]["etag"] == "some_etag"
+    assert results[path]["remote_content"] == "remote_content"
+
+
+@pytest.mark.asyncio
+async def test_backup_max_limit(coordinator, tmp_path):
+    """Test that backups exceeding max_backups are cleaned up."""
+    bp_file = tmp_path / "test.yaml"
+    bp_file.write_text("v0")
+
+    coordinator.config_entry = MagicMock()
+    coordinator.config_entry.options = MappingProxyType({"max_backups": 2})
+    coordinator.hass.async_add_executor_job = AsyncMock(side_effect=lambda fn, *args: fn(*args))
+
+    for i in range(1, 5):
+        await coordinator.async_install_blueprint(
+            str(bp_file), f"v{i}", reload_services=False, backup=True
+        )
+
+    assert bp_file.read_text() == "v4"
+    assert (tmp_path / "test.yaml.bak.1").read_text() == "v3"
+    assert (tmp_path / "test.yaml.bak.2").read_text() == "v2"
+    assert not (tmp_path / "test.yaml.bak.3").exists()
+
+
+@pytest.mark.asyncio
+async def test_backup_migration_old_bak(coordinator, tmp_path):
+    """Test migration of old .bak format to .bak.1 on next backup."""
+    bp_file = tmp_path / "test.yaml"
+    bp_file.write_text("current")
+    (tmp_path / "test.yaml.bak").write_text("old_backup")
+
+    coordinator.config_entry = MagicMock()
+    coordinator.config_entry.options = MappingProxyType({"max_backups": 3})
+    coordinator.hass.async_add_executor_job = AsyncMock(side_effect=lambda fn, *args: fn(*args))
+
+    await coordinator.async_install_blueprint(
+        str(bp_file), "new_version", reload_services=False, backup=True
+    )
+
+    assert bp_file.read_text() == "new_version"
+    assert (tmp_path / "test.yaml.bak.1").read_text() == "current"
+    assert (tmp_path / "test.yaml.bak.2").read_text() == "old_backup"
+    assert not (tmp_path / "test.yaml.bak").exists()
+
+
+@pytest.mark.asyncio
+async def test_backup_rotation(coordinator, tmp_path):
+    """Test that backups rotate correctly: .bak.1 is newest, .bak.3 is oldest."""
+    bp_file = tmp_path / "test.yaml"
+    bp_file.write_text("version_0")
+
+    coordinator.config_entry = MagicMock()
+    coordinator.config_entry.options = MappingProxyType({"max_backups": 3})
+    coordinator.hass.async_add_executor_job = AsyncMock(side_effect=lambda fn, *args: fn(*args))
+
+    for i in range(1, 4):
+        await coordinator.async_install_blueprint(
+            str(bp_file), f"version_{i}", reload_services=False, backup=True
+        )
+
+    assert bp_file.read_text() == "version_3"
+    assert (tmp_path / "test.yaml.bak.1").read_text() == "version_2"
+    assert (tmp_path / "test.yaml.bak.2").read_text() == "version_1"
+    assert (tmp_path / "test.yaml.bak.3").read_text() == "version_0"
+
+
+@pytest.mark.asyncio
+async def test_async_restore_blueprint_error(hass, coordinator):
+    """Test error handling during blueprint restoration."""
+    path = "/config/blueprints/automation/test.yaml"
+    coordinator.data = {path: {"updatable": False}}
+
+    with (
+        patch("builtins.open", MagicMock()),
+        patch("custom_components.blueprints_updater.coordinator.os.path.isfile", return_value=True),
+        patch("custom_components.blueprints_updater.coordinator.os.remove"),
+        patch("custom_components.blueprints_updater.coordinator.os.rename"),
+        patch("custom_components.blueprints_updater.coordinator.shutil.copy2"),
+        patch(
+            "custom_components.blueprints_updater.coordinator.os.replace",
+            side_effect=Exception("Disk error"),
+        ),
+    ):
+        result = await coordinator.async_restore_blueprint(path)
+
+    assert result["success"] is False
+    assert result["translation_key"] == "system_error"
+    assert "Disk error" in result["translation_kwargs"]["error"]
+
+
+@pytest.mark.asyncio
+async def test_async_restore_blueprint_missing(hass, coordinator):
+    """Test restoration when backup is missing."""
+    path = "/config/blueprints/automation/test.yaml"
+    coordinator.data = {path: {"updatable": False}}
+
+    with (
+        patch(
+            "custom_components.blueprints_updater.coordinator.os.path.isfile", return_value=False
+        ),
+    ):
+        result = await coordinator.async_restore_blueprint(path)
+
+    assert result["success"] is False
+    assert result["translation_key"] == "missing_backup"
+
+
+@pytest.mark.asyncio
+async def test_async_restore_blueprint_success(hass, coordinator):
+    """Test successful restoration of a blueprint backup."""
+    path = "/config/blueprints/automation/test.yaml"
+    coordinator.data = {path: {"updatable": False}}
+
+    hass.services.has_service = MagicMock(return_value=True)
+    hass.services.async_call = AsyncMock()
+    coordinator.async_request_refresh = AsyncMock()
+
+    with (
+        patch("builtins.open", MagicMock()),
+        patch(
+            "custom_components.blueprints_updater.coordinator.os.path.realpath",
+            side_effect=os.path.normpath,
+        ),
+        patch("custom_components.blueprints_updater.coordinator.os.path.isfile", return_value=True),
+        patch("custom_components.blueprints_updater.coordinator.os.replace") as mock_replace,
+        patch("custom_components.blueprints_updater.coordinator.os.remove"),
+        patch("custom_components.blueprints_updater.coordinator.os.rename"),
+        patch("custom_components.blueprints_updater.coordinator.shutil.copy2"),
+    ):
+        result = await coordinator.async_restore_blueprint(path)
+
+    mock_replace.assert_any_call(
+        os.path.normpath(f"{path}.bak.1"), os.path.normpath(f"{path}.bak.2")
+    )
+    assert result["success"] is True
+    assert result["translation_key"] == "success"
+    hass.services.async_call.assert_any_call("automation", "reload")
+
+
+@pytest.mark.asyncio
+async def test_async_restore_blueprint_unsafe_path(coordinator):
+    """Test that restoring to an unsafe path is blocked."""
+    coordinator._is_safe_path = BlueprintUpdateCoordinator._is_safe_path.__get__(coordinator)
+    result = await coordinator.async_restore_blueprint("/config/secrets.yaml")
+    assert result["success"] is False
+    assert result["translation_key"] == "system_error"
+
+
+@pytest.mark.asyncio
+async def test_restore_versioned(coordinator, tmp_path):
+    """Test restoring from a specific backup version."""
+    bp_file = tmp_path / "test.yaml"
+    bp_file.write_text("current")
+    (tmp_path / "test.yaml.bak.1").write_text("backup_v1")
+    (tmp_path / "test.yaml.bak.2").write_text("backup_v2")
+
+    coordinator.hass.async_add_executor_job = AsyncMock(side_effect=lambda fn, *args: fn(*args))
+    coordinator.async_reload_services = AsyncMock()
+    coordinator.async_request_refresh = AsyncMock()
+
+    result = await coordinator.async_restore_blueprint(str(bp_file), version=2)
+    assert result["success"] is True
+    assert bp_file.read_text() == "backup_v2"
+    assert (tmp_path / "test.yaml.bak.2").read_text() == "backup_v1"
