+++ tests/ui/test_init_edge_cases.py
