+++ tests/ui/test_translations.py
