+++ tests/utils/test_cdn.py
